requirements.txt
# Core libraries
numpy>=1.21.0
pandas>=2.0.0
scipy>=1.7.0

# Parquet support (cần thiết)
pyarrow>=10.0.0
# hoặc fastparquet>=2022.11.0

# Machine Learning
scikit-learn>=1.0.0
xgboost>=1.5.0
lightgbm>=3.3.0
imbalanced-learn>=0.8.0

# Visualization
matplotlib>=3.4.0
seaborn>=0.11.0
plotly>=5.3.0

# Data processing
pyyaml>=5.4.0
joblib>=1.1.0

# Model interpretation
shap>=0.40.0

# Utilities
tqdm>=4.62.0